Update NPM dependencies
The guy who maintains get-port is a jerk for making his entire army
of NPM packages pure ESM and making everyone else deal with the
fallout. 'get-port' is getting the axe as soon as I don't have
something better to do.

> I plan to migrate most of my (1K+) packages to ESM within 2021.

Yeah, NPM is fine guys, there's nothing wrong with the ecosystem.